moved all course update functions to pastquestions utility file
